turn training api into helper
