+++ mohawk/scripts/mohawk.py
@@ -1,12 +1,8 @@
 #!/usr/bin/env python
-import time
 import os
 import click
 from mohawk.classify import classify as classify_dl
-from mohawk.trainer import trainer
-from mohawk.models import (SmallConvNet, ConvNet2, ConvNetAvg, ConvNetAvg2,
-                           ConvNetAvg3, ConvNetAvg4, ConvNetAvg5, ConvNetAvg6)
-import pandas as pd
+from mohawk.trainer import train_helper
 
 
 @click.group()
@@ -38,15 +34,6 @@ def classify(model_path, sequence_file, output_file, length):
 
 
 # TODO is there a way to do this programmatically?
-model_names_to_obj = {'SmallConvNet': SmallConvNet,
-                      'ConvNet2': ConvNet2,
-                      'ConvNetAvg': ConvNetAvg,
-                      'ConvNetAvg2': ConvNetAvg2,
-                      'ConvNetAvg3': ConvNetAvg3,
-                      'ConvNetAvg4': ConvNetAvg4,
-                      'ConvNetAvg5': ConvNetAvg5,
-                      'ConvNetAvg6': ConvNetAvg6
-                      }
 
 
 @mohawk.command()
@@ -101,7 +88,7 @@ model_names_to_obj = {'SmallConvNet': SmallConvNet,
               help="A tab separated file containing two columns, with the "
                    "name of hyper-parameter in the first column, and the "
                    "value in the second column, and no header.")
-@click.option('--append-time', type=str, default=True, required=False,
+@click.option('--append-time', type=bool, default=True, required=False,
               help="Whether the time should be appended to the supplied "
                    "log_dir argument, which is useful for reducing log_dir "
                    "collision")
@@ -110,69 +97,10 @@ def train(model_name, genome_ids, external_validation_ids, metadata, lr,
           seed, concise_summary, gpu, batch_size, data_dir,
           additional_hyper_parameters, append_time):
     # TODO throw better error
-    start_time = time.time()
-    model = model_names_to_obj[model_name]
-
-    id_list, distribution, classes, n_reads = id_file_loader(genome_ids)
-
-    if external_validation_ids is not None:
-        ext_ids, ext_dist, ext_classes, ext_reads = id_file_loader(
-            external_validation_ids)
-    else:
-        ext_ids, ext_dist, ext_classes, ext_reads = None, None, None, None
-
-    train_kwargs = {'gpu': gpu,
-                    'learning_rate': lr,
-                    'summary_interval': summary_interval,
-                    'epochs': epochs,
-                    'summarize': summarize,
-                    'log_dir': log_dir,
-                    }
-    summary_kwargs = {'concise': concise_summary}
-
-    if additional_hyper_parameters is not None:
-        additional_hparams = parse_hparams_file(additional_hyper_parameters)
-    else:
-        additional_hparams = None
-
-    trainer(model, n_reads, length, train_ratio, id_list=id_list,
-            metadata=metadata, distribution=distribution, class_list=classes,
-            batch_size=batch_size, data_directory=data_dir, random_seed=seed,
-            external_validation_ids=ext_ids,
-            n_external_validation_reads=ext_reads,
-            external_validation_distribution=ext_dist,
-            external_validation_classes=ext_classes,
-            start_time=start_time, train_kwargs=train_kwargs,
-            summary_kwargs=summary_kwargs,
-            additional_hparams=additional_hparams,
-            append_time=append_time)
-
-
-def parse_hparams_file(fp):
-    df = pd.read_csv(fp, sep='\t', index_col=0)
-    dict_ = df.to_dict()['0']
-    out_dict = dict()
-    for key, val in dict_.items():
-        # try to make the key numerical, but if not able to, leave as str
-        try:
-            out_dict[key] = float(val)
-        except KeyError:
-            out_dict[key] = val
-    return out_dict
-
-
-def id_file_loader(genome_ids):
-    id_df = pd.read_csv(genome_ids, sep='\t')
-    id_list = id_df['id']
-    n_reads_by_genome = id_df['n_reads']
-    n_reads = int(n_reads_by_genome.sum())
-    distribution = n_reads_by_genome / n_reads
-    if 'class' in id_df.columns:
-        classes = id_df['class']
-    else:
-        classes = None
-
-    return id_list, distribution, classes, n_reads
+    train_helper(model_name, genome_ids, external_validation_ids, metadata, lr,
+                 epochs, summarize, log_dir, summary_interval, train_ratio,
+                 length, seed, concise_summary, gpu, batch_size, data_dir,
+                 additional_hyper_parameters, append_time)
 
 
 if __name__ == '__main__':
